+++ netbox_idrac_inventory/__init__.py
@@ -1,6 +1,6 @@
 from netbox.plugins import PluginConfig
 
-__version__ = "0.3.4"
+__version__ = "0.3.5"
 
 
 class DellInventoryConfig(PluginConfig):